refactor: admin controllers
